Update DALAnalyzer.java
Pretty sure this should work, after moving through the code "by hand".  Having some trouble fixing some weird parts of my compiler. Basically what it does is enter a new "environment" when you see a '(' and then only exit when the parens are balanced. If the parens are not balanced, then throw an exception on the whole thing and only tag "SEXP"